pmd: fix tracking failure after a while
ICMP and ICMPv6 sequence numbers are 16 bits but are stored in an
unsigned 32 bits variables. After a while (about 10 hours with an
interval of 500ms) pmd sequence number pme->icmp_sequence is over 16
bits. pmd expects a reply with a seq number over 16 bits but the actual
sequence number is on 16 bits.

Old-PR: 71874
Acked-by: Philippe Guibert <philippe.guibert@6wind.com>
(cherry picked from commit 0d113b6a68a4d39111d52c9cb3c731f34eb6ec06)

PR=76618
Acked-by: Louis Scalbert <louis.scalbert@6wind.com>

+++ pmd/pm_echo.h
@@ -52,7 +52,7 @@ struct pm_echo {
 	int echofd_rx_ipv6;
 	uint8_t *tx_buf;
 	uint8_t *rx_buf;
-	uint32_t icmp_sequence;
+	uint16_t icmp_sequence;
 
 	struct thread *t_echo_tmo;
 	struct thread *t_echo_send;
